+++ apps/doc/src/app/version-manager/current.const.ts
@@ -1,1 +1,1 @@
-export const PRIZM_CURRENT_VERSION = '4.6.1';
+export const PRIZM_CURRENT_VERSION = '4.7.0';